+++ src/main/java/dev/dahye/lotto/domain/LottoNumberExtractor.java
@@ -1,7 +1,5 @@
 package dev.dahye.lotto.domain;
 
-import dev.dahye.lotto.util.LottoNumber;
-
 import java.util.ArrayList;
 import java.util.List;
