+++ src/main/java/dev/dahye/lotto/domain/LottoTicket.java
@@ -1,7 +1,5 @@
 package dev.dahye.lotto.domain;
 
-import dev.dahye.lotto.util.LottoNumber;
-
 import java.util.*;
 
 public class LottoTicket {
