+++ src/test/java/dev/dahye/lotto/domain/LottoNumberExtractorTest.java
@@ -1,6 +1,5 @@
 package dev.dahye.lotto.domain;
 
-import dev.dahye.lotto.util.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
