+++ src/test/java/dev/dahye/lotto/util/LottoNumberTest.java
@@ -1,5 +1,6 @@
 package dev.dahye.lotto.util;
 
+import dev.dahye.lotto.domain.LottoNumber;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
